Disambiguate binary signatures by filename too
Binary signatures are not random, but are instead generated from the
build time and mapping size.  A fast build machine can easily build 2
binaries in the same second, which can cause them to both have the same
signature, which would prevent the symbol server from serving them both
up independently.  With this change, the binaries can have different
signatures and so will both be available.